LayoutTests/imported/w3c/web-platform-tests/css/css-scroll-snap/snap-after-relayout/multiple-aligned-targets/prefer-targeted-element-positioned-expected.txt


FAIL prefer-targeted-element-positioned assert_equals: scroller followed target1 in y axis after layout change expected 500 but got 400
FAIL prefer-targeted-element-positioned 1 assert_equals: scroller followed target2 in y axis after layout change expected 500 but got 400
FAIL prefer-targeted-element-positioned 2 assert_equals: scroller followed target3 in y axis after layout change expected 500 but got 400
FAIL prefer-targeted-element-positioned 3 assert_equals: scroller followed target4 in y axis after layout change expected 500 but got 400
FAIL prefer-targeted-element-positioned 4 assert_equals: scroller followed target5 in y axis after layout change expected 500 but got 400
